Null checks for times when no console can be found

+++ src/main/groovy/fi/jasoft/plugin/tasks/RunTask.groovy
@@ -46,8 +46,12 @@ public class RunTask extends DefaultTask {
         if (project.vaadin.plugin.terminateOnEnter) {
             server.start()
             project.logger.lifecycle('Press [Enter] to terminate server...')
-            Util.readLine("")
-            server.terminate()
+            if(Util.readLine("") == null){
+                project.logger.warn("Could not retrieve console. Use [CTRL+C] to terminate server.")
+                server.startAndBlock()
+            } else {
+                server.terminate()
+            }
         } else {
             server.startAndBlock()
         }
